Changes to Hanjie Puzzles
Caused the solutions to appear for each puzzle depending on which puzzle button was selected. Set the initial load up page to have all of the puzzles solutions to be covered by gold, but, if another puzzle is selected, the gold boxes wont cover up the solutions to the puzzles.