Final Commit. Ready for Review
Created alert messages depending on the situation, one to tell the user not to refresh the page or to switch the puzzle since the user will loose their progress, another to tell the user whether the puzzle has been solved or not. We also set a times on the peek, so that it only highlights the incorrect answers for 5 seconds.